Endpoint refatorado para manusear inserções de imagem

+++ src/controllers/productController.js
@@ -15,38 +15,30 @@ const createProduct = async (req, res) => {
 }
 
 const updateProduct = async (req, res) => {
-    const { file } = req
     const { id } = req.params
 
     //verificar se o produto existe
-    const product = await productRepository.findOne({ id: req.params.id })
+    const product = await productRepository.findOne({ id })
 
     if (!product) throw new BadRequestError('Produto não encontrado')
 
+
     //verificar se o produto tem imagem cadastrada, se tiver imagem cadastrada, deletar a imagem antiga
-    if (product.produto_imagem) {
-        const path = product.produto_imagem.split('/').slice(3).join('/')
-        await deleteFile(path)
-    }
+    if (req.body.produto_imagem) {
+        if (product.produto_imagem) {
+            const path = product.produto_imagem.split('/').slice(3).join('/')
+            await deleteFile(path)
+        }
+        await productRepository.update(req.body, { id: req.params.id })
 
-    // verificar se o produto tem imagem nova, se tiver, fazer upload de imagem no servidor de imagem e atualizar o campo produto_imagem
-    if (file) {
-        await uploadFile(uuidv4(), file.buffer, file.mimetype)
-       
-        await productRepository.update({ produto_imagem: file.location }, { id: req.params.id })
+        return res.status(204).send()
     }
+    req.body.produto_imagem = null
 
-    // verificar se o produto novo não tem imagem, e atribuir o valor null ao campo produto_imagem
-    if (!file) {
-        await productRepository.update({ produto_imagem: null }, { id: req.params.id })
-    }
-    // atualizar o produto
-    const updatedProduct = await productRepository.update(req.body, { id: req.params.id })
-    if (!updatedProduct) throw new InternalServerError('Não foi possível atualizar o produto')
+    await productRepository.update(req.body, { id: req.params.id })
 
     return res.status(204).send()
 }
-
 
 const getProduct = async (req, res) => {
     const { categoria_id } = req.query;
@@ -87,7 +79,7 @@ const deleteProduct = async (req, res) => {
     if (product[0].pedidos.length) throw new BadRequestError('Não é possível excluir um produto que está vinculado a um pedido')
 
     if (product.produto_imagem) {
-        const path = product.produto_imagem.split('/').pop()
+        const path = product.produto_imagem.split('/').slice(3).join('/')
         await deleteFile(path)
     }
